Some comments on the Chrome error
It's probably too late, @vrivas, but papers are not the place to
discuss code errors. Either you run it until you reach 10 valid
results, same as in every other browser, or you fix the error. Either
way, errors should be kept out of the paper. Bear that in mind for the
next iteration of the paper, at least.

Besides, unit testing should keep bugs away from your code, so that's
prior to doing any experiment.